requirements.txt
aiogram==3.15.0
aiosqlite==0.20.0
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.32.0
jinja2==3.1.4
python-multipart==0.0.12
aiofiles==24.1.0
